Parameter Thread_IO.1-IO

include S.IO

type +'a t

The type of blocking computation that will produce a value of type 'a

val return : 'a -> 'a t

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

val in_thread : (unit -> 'a) -> 'a t

in_thread f runs f () in a separate thread, returning a blocking computation that will become determined once execution of f is complete.