Module Session__Lift

Lift S.Now to S.Future

module Ident : sig ... end

The identity monad as a stub for blocking computations.

module IO : functor (IO : Session.S.IO) -> functor (Now : Session.S.Now) -> Session.S.Future with type +'a io = 'a IO.t and type t = Now.t and type key = Now.key and type value = Now.value and type period = Now.period

Lift a synchronous S.Now backend to an asynchronous S.Future interface.

module Thread_IO : functor (IO : Session.S.Thread_IO) -> functor (Now : Session.S.Now) -> Session.S.Future with type +'a io = 'a IO.t and type t = Now.t and type key = Now.key and type value = Now.value and type period = Now.period

Lift a synchronous S.Now backend to an asynchronous S.Future interface, using threads. This may add concurrency to the program, depending on how in_thread is implemented.