Module Session__S

Session-related signatures and types.

type error =

| Not_found
The key was not found.

| Not_set
The key was found but had no associated value.

The type of a session error.

These will only be returned by the S.Now.get and S.Future.get operations.

module type Now = sig ... end

The signature for synchronous backends.

module type IO = sig ... end

The signature for a blocking computations.

module type Thread_IO = sig ... end

The signature for blocking computations that can run synchronous computations in a separate thread

module type Future = sig ... end

The signature for asynchronous backends.